Add git-gpush script to simplify pushing patch-sets with reviewers
git-gpush is a wrapper script around a normal 'git push', which
simplifies the way to push patch-sets to Gerrit with reviewers and
CCs.

The script also allows for aliasing, so you can tie Gerrit accounts to
IRC nicks for example. The aliases may be stored in a general file
along side the script, but also as git variables, allowing for repo-
local associations.

The initial .git-gpush-aliases file contains some of the aliases for
use on the qt-labs IRC channel.

Change-Id: I8fcfbd39657ef8aeee6924df28fcf91b4a1fd579
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>