Build documentation while generating source packages.
The documentation is built in a copy of the source, because building
docs currently requires building parts of qtbase, qttools and qtdoc, and
running qmake on every module.  After the docs have been generated they
are copied to qtdoc/doc/html in the clean copy of the sources and the
partially-compiled copy is discarded.

Further changes to other scripts will probably be required to enable
assistant to find the docs in pre-compiled binary packages.

Task-number: QTBUG-26059
Change-Id: I336e62671f3db7cf79d2099ca649e8d4b4b90e0b
Reviewed-by: Simo Fält <simo.falt@nokia.com>