Use bash process substitution to compress simultaneously
First of all, using tar three times is pointless. That scans the entire
source tree three times, adding a lot of disk I/O. At the very least,
this should create one tarball and compress it multiple times.

But we can also compress it multiple times in parallel. A dual-core
machine is enough for this, as xz takes one CPU and bzip2 and gzip will
share the other.

Change-Id: I225aeb0618051a4e24cce654304e0d0de02740c8
Reviewed-by: Simo Fält <simo.falt@nokia.com>