Make qdoc-bot respect work in progress
qdoc-bot should only provide a -1 if something can be staged.
When it is work in progress (WIP or ***) there is no risk of
it getting into Qt, so it should only provide a score of 0
and just tell what the (possible) problems are.

Otherwise it can be misread as a human -1 review to an idea.

Change-Id: Ica2b664eecf375cd0262007a423aafe379c607fd
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@digia.com>
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>