Remove obsolete build configuration for QtWebKit
Building the qtwebkit module works the same way as any other module
these days (i.e. qmake && make), thus using the build-webkit script
for this purpose is obsolete.
In particular when using the qt5_tool the module is actually built
twice: first time with the top-level Makefile, and for the second
time in a final step with build-webkit which fails because the other
build is present. Since the build-webkit script is meant for upstream
developer builds the option can be removed from qt5_tool.
Configure with '-skip qtwebkit' when a webkit build was not requested.

Task-number: QTBUG-30858
Change-Id: Ib6b51b188c9f97064efb834da46de4d03e2a72b7
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@digia.com>