complain about empty lines between footers
they are completely unnecessary and make things look messy.

the algorithm considers cherry-pick lines as footers, following
convention. it even supports multi-line cherry-picks, such as those i
sometimes assemble when squashing multiple commits.

this change will cause a lot of complaints, because almost everybody
appears to be using a broken commit-msg hook. the one coming from our
gerrit is ok, as is the copy in qtrepotools.

Change-Id: I845ba16a70960beb2ce042b2eaddf22b17edcc25
Reviewed-by: Sergio Ahumada <sahumada@blackberry.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>