gpush/gpick: add smart series grouping mode
information from previous pushes is used to deduce the changes which are
part of a series identified by a single change.

the original implementation of this feature would automatically capture
leading loose Changes in front of an existing series, but that turned
out to be a nuisance in practice, because it would often lead to
accidental inclusion of unrelated Changes that were inserted before the
series and forgotten about. we now simply ignore them; a subsequent
change will deal with them.

also, the original implementation used traversal of the actually pushed
commits to reconstruct the series. that approach was dropped in favor of
using dedicated meta data, because the latter is much simpler and also
faster.

Change-Id: Iedb5bd8ca825bcc679b65e524a1c3d76116b58dd
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>