gpush: make the default 'gerrit' remote optional
many projects don't have an authoritative mirror to be used as origin,
so users will clone directly from gerrit. it's pointless to require
them to set up a second remote then.

to this end, determine the upstream remote of the source branch, and use
that when gpush.remote is not configured. the upstream remote itself
will be used by subsequent patches, hence the second option.

Change-Id: I2da0563a7de294b4aa785fc9a01d181e53125438
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>