deal with overzealous garbage collection
every repo should have refs/changes (because fetching changes is the
first thing the bot does after cloning a repo), but with the default
config, gc will pack the refs, and the directory vanishes if no refs
were created for too long (which also breaks our cleanup).
don't error out when that happens.

also document how to avoid that situation to start with.

Change-Id: I92d539785a9810c02263f954d9299169d2fa3f2c
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Simo Fält <simo.falt@digia.com>