gpush: add support for gerrit topic names
ui-wise, i considered the alternative of parsing the topic out of the
target ref like a regular push to old versions of gerrit does. however,
given that specifying a target branch is rather uncommon when using
gpush, this seemed pointless.

the topic is persisted (cached) locally, so the purely off-line --list
mode can make use of it.

Change-Id: I032a1f0c3d2b0e84db594e2e162978a2eff66d21
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>