gpush/gpick: make remembered base always survive the trip through gerrit
gpick in a different clone doesn't have the gpush state file, so it
wouldn't be able to recognize the end of a series which is based on
other pending changes. the solution is to store the base in a custom
field on gerrit.

this obviously needs support in gerrit itself, which is why the feature
is disabled for now. the gerrit support is tracked in QTQAINFRA-1527.

Change-Id: Ia61e70412f56740019093d7aea47174cd7a7065e
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>