move the work-in-progress detection to the bot itself
it obviously makes no sense when the sanitizer is executed locally.

additionally, this allows us a different treatment of WIP reports:
instead of giving a -2 sanity review, give a -2 code review. that way
the change owner is not spammed by useless notifications (because the
sanity review itself is positive, so the message is suppressed by the
gerrit configuration).

Change-Id: I10bc8f8f61673e75983dcb91310836d59c434483
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>