fix email/name complaints not being bound to the commit message
the idea to set the commit-message-indicating "file name" only after
entering the body of the message backfires in REST mode, as there the
author/committer are displayed as part of the commit message.

so instead start out with attributing reports to the commit message, and
interpret a negative line number as an "escape" to the cover message
(used in non-REST mode).

Change-Id: Ib0c810e054ded94aa29768cbc89db10e5df78574
Reviewed-by: Orgad Shaneh <orgads@gmail.com>