Add a clang-format config file
It can be copied or linked to ~/.clang-format or
path/to/qt5/.clang-format

It is based on the WebKit style which is the closest builtin style to Qt.
Then it was adjusted to fit better to the existing Qt style.

The name starts with a '_' instead of with an '.' because I don't want
the file to be hidden in this repository. And clang-format both look for file
named '.clang-format' or '_clang-format'

Change-Id: I0dd561fa9b554331ceaf3a39ea7e3a78649e75c4
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>