git-qt-merge-mainlines: shuffle code to make --help a clean no-op
The call to get_submodules() produced output distracting to those
needing help; move it later and simply set args.modules if unset,
instead of supplying the list as default to the parser.

Moved banner line after arg-parse, too, as it duplicates the help
message's description.  Added an epilog to the help message saying how
to invoke this script.

Change-Id: Ib970f7a9ffa235c9523256d4ccfee448afd37ab2
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>