factor out watches and w-i-p handling to separate worker
originally, the idea was to have one gerrit-bot instance per worker, but
that would result in an (even greater) flood of comments. so all
functionality that didn't fit into sanitize-commit was added to
gerrit-bot. at this time it's getting messy, so clean it up by making it
a plain launcher/dispatcher again, with the ability to run multiple
workers on a single commit.

Change-Id: I8c04f449726793b99e9f1ac5c6db330a1f17389f
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>