gpush: modernize syntax for specifying source and target
initially, git-gpush was just a glorified git-push, and hence it started
out accepting the same command line. however, one of the main
motivations for the script's existence was _avoiding_ the need to
specify things we already know, and therefore the regular push syntax
actually isn't very useful here in the common case.

to that effect, we deprecated the bare remote syntax over two years ago,
in favor of the (very rarely used) --remote option. this allows us now
to interpret a bare word as something more useful: the source of the
push, avoiding the need to always use a colon.

following the same logic, we now introduce the --branch option for
specifying the target branch, expecting that it will not be used much.

these two together allow us to deprecate the old way of specifying both
source and target.

Change-Id: I0964252d911b6e9fe32458cf309c6b5ee4f646ae
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>