gpush: add target branch validation
complain when attempting to push a Change for a branch it hasn't been
pushed for before while it has been pushed for other branches.
add the --force-branch option to override the check.

we specifically ignore merged Changes, to avoid complaining about
cherry-picks. this introduces the risk of creating duplicates for
Changes that were temporarily cherry-picked locally and forgotten about.
this loophole will be addressed by a different mechanism later.

Change-Id: Ifac640a8163ec0f099c53de845fb2c5a372e917a
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>