Automatically move change upon request
Let the gerrit bot listen to comments and process move requests
in the form: "bot: move to 4.5".

This should replace the current method for moving a change, which
is to ask ossi or fregl.

Only the change owner is allowed to request moving the change.

Resume is not needed because the bot is up most of the time, and
unlike the Sanity-Review label, this can be easily retriggered by
the user in case of failure.

Change-Id: I5b89fb8e38b0e0af2b7a4f20a5b29412493fb72e
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>