gpush: add option to capture leading loose Changes
we don't automatically capture leading loose Changes because that would
lead to a significant rate of false positives, but the case of
prepending new Changes to a series is common enough to provide a
shortcut for explicit capture (the "verbose" approach is to simply
regroup the series from scratch).

we complain when encountering leading loose Changes but no capture is
requested, to avoid false negatives in turn. however, until the later
introduction of grouping mode this is somewhat annoying, because the
only ways to make it shut up are to a) capture the loose Changes, b)
push them separately right away, or c) rebase them out of the way.

Change-Id: I128999d06561c24b4f9130b9950106aa142ab64d
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>