Don't check spacing in test-data
We have a general convention of storing test data in testdata/
sub-directories alongside their tests.  This data is whatever it needs
to be, for whatever the test may be doing, so we don't want to waste
cycles commenting on its spacing, or spam reviewers with the results.
Adding this exception might even encourage those maintaining tests
with such data to adopt this (already common) convention and make it
easier for others who encounter tests to know where to find their
data.  (I'll start with testlib's selftest's expected_* data when I
can find the time.)

Change-Id: Ib27b75e858e0573a65540787f4ccae6b2383bc3d
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>