Permit the use of MAC/mac again in commit messages/content
Once in a while somebody runs into the bot complaining about the use of
"MAC" when it's perfectly fine. For example in the context of "MAC
address". This is quite annoying and less helpful than perhaps intended.
So this change removes the trigger-happy part of the check but keeps the
more clearly defined potential uses of old terminology from Apple.

Fixes: QTQAINFRA-2800
Change-Id: If6b9cfbd0a808b424ffc93728fcc66939f5ad6ee
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>