Treat wip branches like dev to suppress cherry-pick messages
wip/* branches don't need to be notified of missing pick-to footers
since users working on wip branches should know what they're doing.
Even though cherry-pick bot is enabled in an entire repo, the
usefulness of sanity-checking for pick-to footers is very limited
in wip branches. The easiest way around this is to treat wip
branches like dev when running in gerrit.

Task-number: QTQAINFRA-3751
Change-Id: I81fd427e83ebcd703f9b1700bf00367544c98a0e
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>