Sanity-bot: Make sure Pick-to footer doesn't contain source branch
The sanity bot should warn if the source branch is in the Pick-to
footer since this would be redundant and cause a failed pick attempt
by the cherry-pick bot. Only exact branches are matched, since
it is possible to submit a change to a minor branch with a Pick-to
footer that targets a bugfix branch.

Fixes: QTQAINFRA-3886
Change-Id: I234d20205bd3896b7be631ba29db2efe895b9bd3
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>