gpick: beautify PatchSet fetching debug output somewhat
don't alias the ref when we actually already have the PatchSet. this
changes only the debug output, as changing the ref would be simply a
noop. we may now try to visit already visited commits, for which we'll
get corresponding debug output as well.

also sort the keys in the surrounding loop.

Change-Id: I096b69f32964038b20bd1b0e8ca4ec1c33f4725d
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>