gpush/gpick: Use resolved Gerrit URL for both pushing and fetching changes
The user might have set up a remote with a separate push URL, e.g.:

    [url "git://code.qt.io/"]
        insteadOf = qt-project:
    [url "ssh://codereview.qt-project.org/"]
        pushInsteadOf = qt-project:

We resolve the Gerrit URL and host in set_gerrit_config(), by taking
insteadOf and pushInsteadOf into account. By using the resolved URL
instead of the remote name when fetching we ensure that we always hit
the Gerrit server. Without this git-gpick will fail.

Change-Id: I580f01b8615c9d9fe5961036b8bb260ef62c8261
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>