Say hello to the new Dependency Update Utility
This replaces the script/bot at qt/qtqa/src/qtmoduleupdater
with a more extensible utility. While this utility can be
plugged into an automation host like Jenkins to act as a bot,
it may also be used as a standalone tool to bring any repo's
dependencies up-to-date with qt/qt5.git or dependency repo HEADs.
See the included README.md for further description of
capabilities, use cases and examples.

Task-number: QTQAINFRA-4594
Change-Id: I6e21f90e410044561bfa4d0dfa72cd09687de321
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>