Have the Submodule Update Utility catch manually merged updates
If a change is manually merged after the utility has marked it as
failed, the utility should pick up on this so the round can
continue automatically without further user action.

This change also updates calls to any() to directly
pass an iterator rather than computing a value set through
list comprehension.

Change-Id: I18a5511de30e92e4276ef596c8ae2011f297dd63
Reviewed-by: Dimitrios Apostolou <jimis@qt.io>