Handle yocto/meta-qt6 LTS branches as public branches
If the script is being run with a tqtc/lts-* branch,
this should be consistent across all repos except for
yocto/meta-qt6 which has no tqtc/ shadow repo and tqtc/lts-*
branches. Handle these branches in the public repo instead.
It can be expected that lts branches in the public repo are
always prefixed as lts-*

This change also includes a drive-by fix to ensure the correct
qt5 repo is updated, taking repo prefixes into account.

Change-Id: I6a8360a6ea682d96882d137b834eea11188c758f
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>