Retry failed staging once in submodule updates
The submodule update bot is not capable of determining
a proper patch-application order for patches to be co-staged
with the submodule update. As a result, some conflicts may be created
due to blindly staging changes. This can be somewhat alleviated
by retrying failed items once before giving up such that
necessary dependencies are staged before the dependent ones.
This brute-force staging method should be suitable for the majority
of cases but may still occasionally result in failed co-staging
when the count of dependent changes is too large.

Change-Id: Idf6bd929eb89b6d6b25f6d0885f225b1a8b83c00
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>