Warn about header files with a common non-.h extension
It is an implied convention in the Qt project that header files should
use a ".h" extension.

As this might non be known by a contributor (especially if the
contributor is from outside the QtCompany) or escape review, possibly
requiring a bigger amount of work to change later, a new warning in the
commit sanitizer script was included to warn about new files that have
some common header extension that is not ".h".

Change-Id: I83a25d829a63811c08a11f2a337707c62423f699
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>