Complain when pick-to footers do not contain the latest stable branch
If a change targets the latest-1 stable branch, but neglects to
include the current latest, the sanity bot should give a -1
warning with a comment. This situation often arises in changes shortly
after qt branches to a new stable branch, as it's easy to forget the
latest branch target has moved forward.

Task-number: QTQAINFRA-5035
Change-Id: I2067c9b1865bd290183d77e46bca2f491708ca2b
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>