Make querying internal coin safe and use configurable hostname
It turns out Requests will throw an exception if the hostname
cannot be resolved. Handle this error gracefully and fall back.

Further, if we failed to query coin for some reason,
automatic staging should only be disabled for blocking
modules being updated since automatic staging is a global
behavior of the bot.

Change-Id: I2fbb744576dd5ec946f9738875b0bc4ac0191874
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>