gpush: don't confirm specified base with gerrit
the only reasonable case of specifying a non-upstream base commit is
chaining local series. but for these we already know whether gerrit
knows the commit, as we record what has been pushed.

as a consequence, we will now reject attempts to base on older PatchSets
of a given Change. this is arguably a feature.

Change-Id: Ib698e1cc7c05efeb03eb19983257cb6cc413f6e2
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>