gpush: make --group preserve pending properties
second step towards making --all compatible with --group.

--group resetting the pending properties seemed reasonable at the time,
but as more persistent properties are added, this becomes increasingly
annoying. also, it's inconsistent with "committed" (already pushed out)
properties being preserved.

this is a backwards-incompatible behavior change, but this is likely to
be actually a good thing, as the old behavior was unexpectedly
destructive.

this also introduces the --reset-props option which has the behavior
plain --group had previously. the utility of this is debatable, as it
maintains the inconsistent behavior, but whatever.

Change-Id: I35162138e34c885ef200a315a797f531d2c079a5
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>