gpick: try to deduce remote series base from topic
in lieu of gerrit actually saving the base of a series for us, consider
inconsistent topics to be a marker for a series ending before hitting
upstream, and therefore being based on another pending series.

gpush gets a complementary check that the chained series have different
topics. an alternative would be auto-generated topics, but that wouldn't
be too friendly to reviewers.

as a drive-by, we don't try to aggregate the 'hidden' property on loose
Changes any more - these have no persistent properties, so it was
pointless.

Change-Id: I2c538fe91d9c50fcf87fc6284f621fb3e633b2df
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>