gpush/gpick: try to preserve group ids
a subsequent commit will require the gids to stay stable.

it also improves the optics, making debugging somewhat easier:
  git diff refs/gpush/state@{1}..refs/gpush/state
now returns something that actually looks sane in most cases.

if we actually versioned the state file, this would also save space, but
as there is no need for versioning beyond what the reflog provides, this
is kinda moot.

Change-Id: If222c4763527bcfe3629bc181339176de095db39
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>