gpush: introduce infra for displaying additional group meta-data
print additional info between a series' caption and list of Changes.
this will be used for info that is (potentially) way too big to fit into
the caption itself.

to be able to suppress this, we make --list compatible with --quiet.
note that in --group and push modes this is not possible, as there
--quiet just completely suppresses the listing. this seems fine, as
these don't have the different levels of --list/-rebase/-online, either.

as a drive-by, --list becomes compatible with --verbose as well - there
isn't really a reason for it not to be, and since the recent
introduction of --list-rebase it is actually somewhat useful.

Change-Id: I5b4f9a91b73ea4816ac9631d2662765aaa3df7df
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>