gpush/gpick: resolve branch names internally
we have the local refs loaded up, so we can resolve them in-process,
just like we already did with HEAD.
this is an optimization, but the real intent is a prospective bug fix
(if we manipulated the refs only in memory (due to --dry-run) prior to
using them). consequently, we don't do ancestry traversal (e.g., HEAD~3)
in-process (yet).

Change-Id: I1930c61c986d846b37f02ca00aaa5a6c6a5d9bac
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>