Allow non-numeric branches which do not begin with wip/
Some repos are better suited to tracking remote branch names
from other projects, rather than following the Qt branching
scheme. When these branches do not conform to \d+.\d+.\d+,
they are not included in the %allHeads of the sanity bot.
As such, attempting to cherry-pick to any of these non-conforming
branches will be given a -2 by the bot since it is unaware of the
branch, even though the pick target is valid.

Provide an option per repo to bypass, allowing non-numeric branches.

Fixes: QTQAINFRA-5449
Change-Id: I6ca7772703572fdcc010f31dc9bacbbf4fa5e6a2
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>