gpush/gpick: introduce infra for indirect properties
rather than having the same (potentially big) value in each Change it
was assigned to, duplicate only a small id.

an arguably cleaner way to approach this would be saving actual group
properties rather than aggregating them on the fly from Change
properties. but when i implemented this, it turned out that we'd have to
be able to resolve conflicts resulting from merging groups in rather
inopportune places: gpick calls assign_series() from deduce_series()
(though a conflict occurring due to this is a rather academical case)
and from _source_map_finish_initial() via analyze_local_branch(). this
means that we'd have to add the property-overriding command line options
to gpick (which is ugly) or we'd have to croak in gpick and tell the
user to use gpush first instead (which is also ugly).

more broadly, group properties aren't a very good fit for gerrit, where
groups aren't even a thing. this is why we don't use them in the first
place.

Change-Id: I555360de0fc08b698213e56045ca408c1ee47851
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>