gpush: delay (re-)setting the push base to the local branch's base
when the series isn't pushed right away, but only (re-)grouped, then it
is more intuitive to postpone resetting the base until the actual push,
lest one will have to rebase again (when the local series receives
another conflicted rebase in between).

the wording of the help text is intentionally very broad, as it will
soon cover more cases.

Change-Id: I3ca99bc5d011b546e9f014c176c3c1bbc0f7a8e7
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>