gpick: fix dropping closed Changes when reconstructing pushed series
so far, we filtered the remote and the pushed series independently. this
would have the effect that if the two series disagreed about the
presence of Changes in particular states of "closedness", they would be
filtered differently. this would then lead to spurious divergence when
comparing series structure. fix this by carrying over the "threshold of
closedness" from the remote to the pushed series.

we now also drop closed Changes which are still in the local series from
the pushed series, as we do for the remote one. it doesn't really matter
either way (it decides whether the divergence will be first shown in the
local or the remote series, which is irrelevant), so just do the simpler
thing.

Change-Id: I3a61951a10a82eb9548d529281a74fe76850d383
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>