gpick: remove local base hinting when assembling remote series
the idea was to augment the remote series with a possibly present local
base (when series are chained), in case gerrit doesn't save it for us
(which currently is always, as the feature isn't even implemented yet).

however, this had three problems:
- using sha1s is a non-starter for remote series that have been rebased.
- due to missing integration with set_base_deduced(),
  do_adjust_changes() would be unaware of the base.
- the remote side may have re-grouped the series, at which point local
  hinting would make quite a mess.

the first two problems would be fixable with a Change-Id based
implementation in set_base_deduced(), but the last one is inherent.

so just rip it out. we'll introduce topic-based series splitting in a
later commit, which avoids anomalies and is actually visible on gerrit.

Change-Id: I5d56f0498e190ac51d2000f89b570f55e48c7ef7
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>