gpush/gpick: reshuffle sanity check in analyze_local_branch() somewhat
run it on the re-enumerated commits rather than the raw ones. most
visibly, this saves one loop. it also has the effect of checking only
along the 1st parents, which is consistent with the overall operation of
the scripts. another effect will manifest only after a subsequent patch:
no double-examination of some commits.

Change-Id: I60b612a52b9e86470f7b75241e45dae4e8637e77
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>