gpick: revamp tails of advance_*_series()
the pushed/picked commit is already resolved by map_queries(), and with
it comes the base assigned by set_base_{imposed,deduced}(). using this
allows us to remove two indirections and conditionals. more importantly,
the processed data becomes authoritative, which a subsequent commit will
make use of.

this doesn't change anything for remote series; the base fallback chain
which this would have broken was removed in a prior commit.

for pushed series things are more complicated.
firstly, the following affects only Changes that have been rebased out
of the local branch, as the remaining ones are in the initial set and
are therefore handled further up.
when gerrit imposes no base (that is, currently always) we would now set
the deduced base, while formerly we'd have inherited the one from the
child commit. to avoid a regression (ignoring the separation of chained
series), we now fall back to imposing the locally recorded push base
instead.

Change-Id: I1703f614b6571ee7adf6a9452b8f7a404d3b30f1
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>