gpush: introduce "mail" mode
i work on the linux kernel and git, both of which still use patch-based
workflows, at least for non-maintainer contributions. there are various
other tools which support the "stacked branches" workflow gpush centers
around, but none of them is nearly as comfortable as gpush. so there we
go: a gerrit-free mode in "git-gerrit-push".

Change-Id: Ifb2d603a06d0b7825392a37967a5f6c28ed79455
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>