Remove warning about leftover Pick-to footers in cherry-picks
The cherry-pick bot will now "waterfall" changes down from newest to
oldest, so Pick-to footers are expected to remain in cherry-picks until
all targets are exhausted.

Ideally, we'd ensure that no excess entries remain, but that would
require examining the named branches, which would be complex and would
require ensuring that the refs are current, which both seem
unreasonable. Therefore, we simply suppress the check in cherry-picks.

Change-Id: I20a77603779654d113e80121b50491239c048e36
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>