Limit deprecation adjustments grumble
Where developers deprecate things we want to remind them that they
have a duty to sort out all existing Qt code that exercises the
newly-deprecated API. However, we were producing the same warning also
for the #if-ery around blocks of deprecations (which doesn't need it,
as the deprecations within will get it) and for the additions of new
version-specific deprecation defines. These can be distinguished as
preprocessor directives - specifically #if and #define, although #else
or #endif may also have a comment on it echoing #if's condition.

So skip this check on all preprocessor commands, so we only grumble
where the macros are used to deprecate declarations.

Change-Id: Ie296d81a932bd098c54f9e166f3b0609c4af34fd
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>