gpush: cache rebased commits more aggressively
we now save the rebased commits even if they are not actually pushed
out, which is the case with --list-online, and even on a regular push
attempt that errors out due to failing to rebase another series. this
vastly accelerates repeated listing and the subsequent push for series
that are not rebased meanwhile.

Change-Id: I1c9df94193a07d827d7cfea5a51980ff3f47f58a
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>