Set the target branch on missing yocto dependencies
The yocto supermodule update may contain some repos that are not part
of the common update round. If this is the case, then we need to make
sure the target branch is set on the repo following the discovery
process. Otherwise, the default dev or master branch will be chosen.

Change-Id: Id5d5acbab16cccf5fcbba7a8f5d925979da34a8d
Reviewed-by: Daniel Smith <daniel.smith@qt.io>