Ignore branch validity for LTS release branch targets in public repos
Since LTS release branches may not exist in the public repo,
sanitize-commit cannot validate Pick-to: targets for them when
evaluating public commits. This change bypasses the branch
validation check for release targets to known LTS branches.

In the event that the target branch is not a valid branch in the end,
the cherry-pick bot will simply fail to perform the requested pick
and post an error message to the change, alerting the author.

Fixes: QTQAINFRA-6191
Change-Id: Id87b9e6fd379ad741801f339d42c13dca1f51123
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>