gpush/gpick: handle Changes w/o target branch on disappeared branches
in source_map_traverse(), we try to find Changes in the upstream of
their target branch, in case they were already integrated. but Changes
that were never pushed (and weren't --group'd with --branch) don't have
a target branch, so we'd get complaints about indexing by undef. we
should just filter them out, as they don't have a base commit, either,
and will therefore be skipped by the subsequent loop we're collecting
the data for, too.

Change-Id: Ic8fea075ff0b8cb6cb44db3353a9cdb4d52e276b
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>